Move entityAttributeText to data-utils.js
